# docs/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ErrorResponse" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "details" : {
            "items" : {
              "$ref" : "#/components/schemas/FieldIssue"
            },
            "type" : "array"
          },
          "message" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "FieldIssue" : {
        "properties" : {
          "field" : {
            "type" : "string"
          },
          "issue" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "IocDto" : {
        "properties" : {
          "attribution" : {
            "items" : {
              "$ref" : "#/components/schemas/AttributionDto"
            },
            "type" : "array"
          },
          "confidence" : {
            "format" : "int32",
            "type" : "integer"
          },
          "firstSeen" : {
            "format" : "date-time",
            "type" : "string"
          },
          "hashType" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "lastSeen" : {
            "format" : "date-time",
            "type" : "string"
          },
          "score" : {
            "format" : "int32",
            "type" : "integer"
          },
          "severity" : {
            "type" : "string"
          },
          "sourceCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "status" : {
            "type" : "string"
          },
          "tags" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array",
            "uniqueItems" : true
          },
          "tlp" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          },
          "validUntil" : {
            "format" : "date-time",
            "type" : "string"
          },
          "value" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "IocSourceDto" : {
        "properties" : {
          "redistributionPolicy" : {
            "type" : "string"
          },
          "reportCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sourceConfidence" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sourceFirstSeen" : {
            "format" : "date-time",
            "type" : "string"
          },
          "sourceId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "sourceLastSeen" : {
            "format" : "date-time",
            "type" : "string"
          },
          "sourceName" : {
            "type" : "string"
          },
          "sourceSeverity" : {
            "type" : "string"
          },
          "sourceTlp" : {
            "type" : "string"
          },
          "sourceValidUntil" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "LookupRequest" : {
        "properties" : {
          "values" : {
            "items" : {
              "type" : "string"
            },
            "minItems" : 1,
            "type" : "array"
          }
        },
        "required" : [ "values" ],
        "type" : "object"
      },
      "LookupResponse" : {
        "properties" : {
          "results" : {
            "items" : {
              "$ref" : "#/components/schemas/Result"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "PageResponse" : {
        "properties" : {
          "hasMore" : {
            "type" : "boolean"
          },
          "items" : {
            "items" : { },
            "type" : "array"
          },
          "nextCursor" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "SearchRequest" : {
        "properties" : {
          "confidenceMax" : {
            "format" : "int32",
            "type" : "integer"
          },
          "confidenceMin" : {
            "format" : "int32",
            "type" : "integer"
          },
          "cursor" : {
            "type" : "string"
          },
          "includeExpired" : {
            "type" : "boolean"
          },
          "lastSeenFrom" : {
            "format" : "date-time",
            "type" : "string"
          },
          "lastSeenTo" : {
            "format" : "date-time",
            "type" : "string"
          },
          "limit" : {
            "format" : "int32",
            "type" : "integer"
          },
          "query" : {
            "maxLength" : 2048,
            "minLength" : 0,
            "type" : "string"
          },
          "scoreMax" : {
            "format" : "int32",
            "type" : "integer"
          },
          "scoreMin" : {
            "format" : "int32",
            "type" : "integer"
          },
          "severity" : {
            "type" : "string"
          },
          "sourceId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "tags" : {
            "items" : {
              "maxLength" : 128,
              "minLength" : 0,
              "type" : "string"
            },
            "type" : "array"
          },
          "tlp" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        },
        "required" : [ "query" ],
        "type" : "object"
      },
...
      "SourceStatsDto" : {
        "properties" : {
          "displayName" : {
            "type" : "string"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "indicatorCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "lastSuccessAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "sourceId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "sourceType" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "apiKeyAuth" : {
        "description" : "Machine-to-machine key: ctip_<env>_<32 base62>",
        "in" : "header",
        "name" : "X-API-Key",
        "type" : "apiKey"
      },
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Access token from POST /api/v1/auth/login",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Cyber Threat Intelligence Platform read API. M1 exposes anonymous read access to public TLP:CLEAR intelligence with cursor pagination; authentication, RBAC and write endpoints arrive in M2. All errors share a single ErrorResponse shape with a log-correlated traceId (see docs/spec/09-api.md §9.4).",
    "title" : "CTIP API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/iocs" : {
      "get" : {
        "description" : "Lists visible indicators ordered by (lastSeen DESC, id DESC). Filters: type, severity, status, tlp, tags (repeat the parameter; all must match), sourceId, confidenceMin/Max, scoreMin/Max, lastSeenFrom/To (ISO-8601); EXPIRED entries are excluded unless includeExpired=true. limit above the plan maximum is clamped, not rejected. offset mode is for page-number UIs only and is capped at 10000. 認證:匿名(僅 public TLP:CLEAR)。",
        "operationId" : "list_1",
        "parameters" : [ {
          "in" : "query",
          "name" : "cursor",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "offset",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "type",
          "required" : false,
          "schema" : {
            "enum" : [ "IPV4", "IPV6", "DOMAIN", "URL", "FILE_HASH", "EMAIL" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "severity",
          "required" : false,
          "schema" : {
            "enum" : [ "INFO", "LOW", "MEDIUM", "HIGH", "CRITICAL" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "enum" : [ "ACTIVE", "EXPIRED", "REVOKED", "FALSE_POSITIVE" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "tlp",
          "required" : false,
          "schema" : {
            "enum" : [ "CLEAR", "GREEN", "AMBER", "AMBER_STRICT", "RED" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "includeExpired",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "tags",
          "required" : false,
          "schema" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        }, {
          "in" : "query",
          "name" : "sourceId",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "confidenceMin",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "confidenceMax",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "scoreMin",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "scoreMax",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "lastSeenFrom",
          "required" : false,
          "schema" : {
            "format" : "date-time",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "lastSeenTo",
          "required" : false,
          "schema" : {
            "format" : "date-time",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "example" : {
                  "hasMore" : true,
                  "items" : [ {
                    "attribution" : [ {
                      "homepage" : null,
                      "sourceName" : "Mock OpenPhish"
                    } ],
                    "confidence" : 60,
                    "firstSeen" : "2026-08-01T00:00:00Z",
                    "hashType" : null,
                    "id" : "1f0d2c4e-93a5-4f6b-8c1d-2e3a4b5c6d7e",
                    "lastSeen" : "2026-08-20T10:00:00Z",
                    "score" : 42,
                    "severity" : "HIGH",
                    "sourceCount" : 1,
                    "status" : "ACTIVE",
                    "tags" : [ "sample", "phishing" ],
                    "tlp" : "CLEAR",
                    "type" : "DOMAIN",
                    "validUntil" : "2026-11-18T10:00:00Z",
                    "value" : "mal-8.ctip-sample.net"
                  } ],
                  "nextCursor" : "eyJscyI6IjIwMjYtMDgtMjBUMTA6MDA6MDBaIiwiaWQiOiIuLi4ifQ=="
                },
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponse"
                }
              }
            },
            "description" : "One page of indicators"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "INVALID_CURSOR / OFFSET_TOO_LARGE / INVALID_REQUEST"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit exceeded (RATE_LIMIT_EXCEEDED)"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected error (INTERNAL_ERROR)"
          }
        },
        "security" : [ ],
        "summary" : "List IOCs (cursor pagination)",
        "tags" : [ "IOC" ]
      }
    },
    "/api/v1/iocs/lookup" : {
      "post" : {
        "description" : "Verifies up to the plan limit of values in one call (anonymous: 20). Values are cleaned and normalized before identity matching; unparsable or invisible values are reported as found=false, never as errors. Bloom-filter clients must confirm hits here (11 §11.6). 認證:匿名。",
        "operationId" : "lookup",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "example" : {
                "values" : [ "MAL-8.CTIP-SAMPLE.NET.", "203.0.113.7" ]
              },
              "schema" : {
                "$ref" : "#/components/schemas/LookupRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "example" : {
                  "results" : [ {
                    "found" : true,
                    "ioc" : {
                      "attribution" : [ {
                        "homepage" : null,
                        "sourceName" : "Mock OpenPhish"
                      } ],
                      "confidence" : 60,
                      "firstSeen" : "2026-08-01T00:00:00Z",
                      "hashType" : null,
                      "id" : "1f0d2c4e-93a5-4f6b-8c1d-2e3a4b5c6d7e",
                      "lastSeen" : "2026-08-20T10:00:00Z",
                      "score" : 42,
                      "severity" : "HIGH",
                      "sourceCount" : 1,
                      "status" : "ACTIVE",
                      "tags" : [ "sample", "phishing" ],
                      "tlp" : "CLEAR",
                      "type" : "DOMAIN",
                      "validUntil" : "2026-11-18T10:00:00Z",
                      "value" : "mal-8.ctip-sample.net"
                    },
                    "value" : "MAL-8.CTIP-SAMPLE.NET."
                  }, {
                    "found" : false,
                    "ioc" : null,
                    "value" : "203.0.113.7"
                  } ]
                },
                "schema" : {
                  "$ref" : "#/components/schemas/LookupResponse"
                }
              }
            },
            "description" : "Per-value verification results"
          },
          "413" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "PAYLOAD_TOO_LARGE (batch above plan limit)"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit exceeded (RATE_LIMIT_EXCEEDED)"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected error (INTERNAL_ERROR)"
          }
        },
        "security" : [ ],
        "summary" : "Batch exact lookup",
        "tags" : [ "IOC" ]
      }
    },
    "/api/v1/iocs/search" : {
      "post" : {
        "description" : "Substring search over the canonical (normalized) value with the same filters and cursor pagination as the list endpoint (type, severity, status, tlp, tags, sourceId, confidence/score ranges, lastSeen range). M1 backs this with PostgreSQL (pg_trgm / GIN indexes); M2 swaps in Elasticsearch behind the same contract. 認證:匿名。",
        "operationId" : "search",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "example" : {
                "lastSeenFrom" : "2026-08-01T00:00:00Z",
                "limit" : 5,
                "query" : "ctip-sample",
                "scoreMin" : 20,
                "tags" : [ "phishing" ],
                "type" : "DOMAIN"
              },
              "schema" : {
                "$ref" : "#/components/schemas/SearchRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "example" : {
                  "hasMore" : true,
                  "items" : [ {
                    "attribution" : [ {
                      "homepage" : null,
                      "sourceName" : "Mock OpenPhish"
                    } ],
                    "confidence" : 60,
                    "firstSeen" : "2026-08-01T00:00:00Z",
                    "hashType" : null,
                    "id" : "1f0d2c4e-93a5-4f6b-8c1d-2e3a4b5c6d7e",
                    "lastSeen" : "2026-08-20T10:00:00Z",
                    "score" : 42,
                    "severity" : "HIGH",
                    "sourceCount" : 1,
                    "status" : "ACTIVE",
                    "tags" : [ "sample", "phishing" ],
                    "tlp" : "CLEAR",
                    "type" : "DOMAIN",
                    "validUntil" : "2026-11-18T10:00:00Z",
                    "value" : "mal-8.ctip-sample.net"
                  } ],
                  "nextCursor" : "eyJscyI6IjIwMjYtMDgtMjBUMTA6MDA6MDBaIiwiaWQiOiIuLi4ifQ=="
                },
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponse"
                }
              }
            },
            "description" : "One page of matches"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "INVALID_REQUEST (blank query, unknown enum value, bad cursor)"
          },
          "415" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "UNSUPPORTED_MEDIA_TYPE"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit exceeded (RATE_LIMIT_EXCEEDED)"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected error (INTERNAL_ERROR)"
          }
        },
        "security" : [ ],
        "summary" : "Search IOCs",
        "tags" : [ "IOC" ]
      }
    },
    "/api/v1/iocs/{id}" : {
      "get" : {
        "description" : "Returns a single indicator by id. Cross-tenant or invisible resources are always 404 (existence is never disclosed). 認證:匿名。",
        "operationId" : "byId_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "example" : {
                  "attribution" : [ {
                    "homepage" : null,
                    "sourceName" : "Mock OpenPhish"
                  } ],
                  "confidence" : 60,
                  "firstSeen" : "2026-08-01T00:00:00Z",
                  "hashType" : null,
                  "id" : "1f0d2c4e-93a5-4f6b-8c1d-2e3a4b5c6d7e",
                  "lastSeen" : "2026-08-20T10:00:00Z",
                  "score" : 42,
                  "severity" : "HIGH",
                  "sourceCount" : 1,
                  "status" : "ACTIVE",
                  "tags" : [ "sample", "phishing" ],
                  "tlp" : "CLEAR",
                  "type" : "DOMAIN",
                  "validUntil" : "2026-11-18T10:00:00Z",
                  "value" : "mal-8.ctip-sample.net"
                },
                "schema" : {
                  "$ref" : "#/components/schemas/IocDto"
                }
              }
            },
            "description" : "The indicator"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "NOT_FOUND (missing or not visible)"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit exceeded (RATE_LIMIT_EXCEEDED)"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected error (INTERNAL_ERROR)"
          }
        },
        "security" : [ ],
        "summary" : "Get one IOC",
        "tags" : [ "IOC" ]
      }
    },
    "/api/v1/iocs/{id}/sources" : {
      "get" : {
        "description" : "Per-source observations for one indicator. Redistribution policy is enforced: for non-owner viewers only PUBLIC_REDISTRIBUTABLE and ATTRIBUTION_REQUIRED records are returned (DERIVED_ONLY and INTERNAL_ONLY are masked). 認證:匿名。",
        "operationId" : "sources_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "example" : [ {
                  "redistributionPolicy" : "ATTRIBUTION_REQUIRED",
                  "reportCount" : 2,
                  "sourceConfidence" : 60,
                  "sourceFirstSeen" : "2026-08-01T00:00:00Z",
                  "sourceId" : "6f0d2c4e-93a5-4f6b-8c1d-2e3a4b5c6d7e",
                  "sourceLastSeen" : "2026-08-20T10:00:00Z",
                  "sourceName" : "Mock OpenPhish",
                  "sourceSeverity" : "HIGH",
                  "sourceTlp" : "CLEAR",
                  "sourceValidUntil" : null,
                  "status" : "ACTIVE"
                } ],
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/IocSourceDto"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Visible source records (may be empty when policy masks them)"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "NOT_FOUND (missing or not visible)"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit exceeded (RATE_LIMIT_EXCEEDED)"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected error (INTERNAL_ERROR)"
          }
        },
        "security" : [ ],
        "summary" : "Get IOC source records",
        "tags" : [ "IOC" ]
      }
    },
...
    "/api/v1/stats/sources" : {
      "get" : {
        "description" : "Indicator counts, enablement, health status and last success time per threat source. 認證:匿名。",
        "operationId" : "sources",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "example" : [ {
                  "displayName" : "Mock OpenPhish",
                  "enabled" : true,
                  "indicatorCount" : 351,
                  "lastSuccessAt" : "2026-08-26T02:00:00Z",
                  "sourceId" : "6f0d2c4e-93a5-4f6b-8c1d-2e3a4b5c6d7e",
                  "sourceType" : "MOCK_OPENPHISH",
                  "status" : "ACTIVE"
                } ],
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/SourceStatsDto"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "One row per configured source"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit exceeded (RATE_LIMIT_EXCEEDED)"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected error (INTERNAL_ERROR)"
          }
        },
        "security" : [ ],
        "summary" : "Per-source counts and health",
        "tags" : [ "Stats" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ],
...
}